+++ src/index.js
@@ -2,5 +2,5 @@ export {getAuthData} from './utils/oauth';
 export * from './useOauthData';
 export * from './utils/getUserInfo';
 export * from './utils/getAccessToken';
-export * from './WithTokensExpirationAccess';
+export * from './withTokensExpirationAccess';
 export {default} from './Provider';