+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- withTokenExpirationAccess to logout user if token is expired.
+
 ### [v.1.5.0] - 2025-02-13
 
 ### Changed